requirements.txt
# Core
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.6,<3
PyYAML>=6,<7
filelock>=3.12,<4

# Tests & tooling (opcionales pero recomendados)
httpx>=0.27,<1.0
pytest>=8,<9
pytest-asyncio>=0.23,<1
